extract asyncTryCatchAssign to new file
