+++ backend/src/user/types.ts
@@ -1,4 +1,4 @@
 import { Profile } from "passport";
-import { TypeLike } from "../types";
+import { TypeLike } from "../shared/types";
 
 export type profileEssentials = TypeLike<Pick<Profile, "displayName" | "emails" | "photos">>;
